将vo,po放入对应的包中
