+++ code-server/src/main/java/com/clsaa/dop/server/code/model/bo/project/TagBo.java
@@ -1,4 +1,4 @@
-package com.clsaa.dop.server.code.model.bo;
+package com.clsaa.dop.server.code.model.bo.project;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -10,9 +10,6 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class TokenBo {
-
-    private String access_token;
-
-
+public class TagBo {
+    String name;
 }
