+++ code-server/src/main/java/com/clsaa/dop/server/code/model/bo/user/TokenBo.java
@@ -1,4 +1,4 @@
-package com.clsaa.dop.server.code.model.bo;
+package com.clsaa.dop.server.code.model.bo.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -10,7 +10,9 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class BranchBo {
+public class TokenBo {
 
-    private String name;
+    private String access_token;
+
+
 }
