+++ code-server/src/main/java/com/clsaa/dop/server/code/service/ProjectService.java
@@ -1,12 +1,10 @@
 package com.clsaa.dop.server.code.service;
 
 
-import com.clsaa.dop.server.code.model.bo.BranchBo;
-import com.clsaa.dop.server.code.model.bo.ProjectBo;
-import com.clsaa.dop.server.code.model.bo.ProjectListBo;
-import com.clsaa.dop.server.code.model.bo.TagBo;
-import com.clsaa.dop.server.code.model.vo.ProjectVo;
-import com.clsaa.dop.server.code.util.BeanUtils;
+import com.clsaa.dop.server.code.model.bo.project.BranchBo;
+import com.clsaa.dop.server.code.model.bo.project.ProjectBo;
+import com.clsaa.dop.server.code.model.bo.project.ProjectListBo;
+import com.clsaa.dop.server.code.model.bo.project.TagBo;
 import com.clsaa.dop.server.code.util.RequestUtil;
 import org.apache.http.NameValuePair;
 import org.apache.http.message.BasicNameValuePair;
@@ -116,12 +114,5 @@ public class ProjectService {
 
     }
 
-    public static void main(String[] args) {
-        ProjectService ps=new ProjectService();
-        ProjectBo projectBo=ps.findProject(3);
-        ProjectVo projectVo=BeanUtils.convertType(projectBo,ProjectVo.class);
-
-        System.out.println(projectVo);
-    }
 
 }
